Add mongoose id verification to functions

+++ service/comment.services.js
@@ -1,6 +1,9 @@
 const Comment = require("../models/Comment");
+const { checkValidObjectID } = require("./mongoose.services");
 
 const findCommentById = async (id) => {
+  const idCheck = checkValidObjectID(id);
+  if (!idCheck) return;
   return Comment.findById(id)
     .lean()
     .populate("author", "_id username userImgKey")
@@ -8,6 +11,9 @@ const findCommentById = async (id) => {
 };
 
 const findPostComments = async (page, limit, postId) => {
+  const idCheck = checkValidObjectID(postId);
+  if (!idCheck) return;
+
   if (page || limit) {
     const pageInt = parseInt(page) || 1;
     const limitInt = parseInt(limit) || 10;
@@ -31,6 +37,10 @@ const findPostComments = async (page, limit, postId) => {
 };
 
 const createNewComment = async (author, parentPostId, commentBody) => {
+  const postIdCheck = checkValidObjectID(parentPostId);
+  const authorIdCheck = checkValidObjectID(author);
+  if (!postIdCheck || !authorIdCheck) return;
+
   return Comment.create({
     author,
     parentPostId,
@@ -39,10 +49,14 @@ const createNewComment = async (author, parentPostId, commentBody) => {
 };
 
 const findAndDeleteComment = async (id) => {
+  const idCheck = checkValidObjectID(id);
+  if (!idCheck) return;
   return Comment.findByIdAndDelete(id).exec();
 };
 
 const countComments = async (parentPostId) => {
+  const idCheck = checkValidObjectID(parentPostId);
+  if (!idCheck) return;
   return Comment.countDocuments({ parentPostId });
 };
